Split echo task out
+ Move echo task to separate file.
+ Fix requires.
